+++ samples/Pkcs12Connect/src/main/java/pkcs12connect/Pkcs12Connect.java
@@ -74,7 +74,7 @@ public class Pkcs12Connect {
             builder.withConnectionEventCallbacks(callbacks)
                 .withClientId(cmdData.input_clientId)
                 .withEndpoint(cmdData.input_endpoint)
-                .withPort((short)cmdData.input_port)
+                .withPort(cmdData.input_port)
                 .withCleanSession(true)
                 .withProtocolOperationTimeoutMs(60000);
             MqttClientConnection connection = builder.build();
